requirements.txt
Django~=4.2.20
djangorestframework~=3.16.0
pip~=23.2.1
wheel~=0.41.2
PyYAML~=6.0.2
attrs~=25.3.0
asgiref~=3.8.1
sqlparse~=0.5.3
serializers~=0.2.4
referencing~=0.36.2
jsonschema~=4.23.0
jsonschema-specifications~=2025.4.1
setuptools~=68.2.0
future~=1.0.0
uritemplate~=4.1.1
inflection~=0.5.1
drf-spectacular~=0.28.0
psycopg2-binary>=2.9